Found a bug in the formalization.
Somehow, our vars %block does not cover everything.
I added a couple of operations and instances,
and preservation broke. We need to fix this first.

Note that all possible cases for eq-inst-op cause
numerous coverage errors, so it may be wise to
comment out some of them when fixing up the
proof of preservation.